Heavy ChainBuilder updates. Changed to single-button page. Fixed package.json.
ChainBuilder now handles one data buffer, and allows a person to reference offsets in that buffer by an id. This can be used to pass in data such as file paths, to allocate an area for storing chain results, and to refer to those offsets in JS to get the result.

Two functions currently allow for passing in data ids, storeR3Data() and syscallR3Data().

Updated ArrayLeaker to be more thorough during getStringAddress(), until more information is known about the structures.

Added some PS3Xploit and original offsets.